# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Motolegacy",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://KidziBack.squaretech.tech/motolegacy/api"
        },
        {
            "url": "http://127.0.0.1:8000/motolegacy/api"
        }
    ],
    "paths": {
...
        "/seller/incoming-purchase-requests": {
            "get": {
                "tags": [
                    "Seller - Requests"
                ],
                "summary": "List purchase requests for seller's cars",
                "operationId": "8551a68254c7436df5af2b8f5db38f1e",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/seller/purchase-requests/{id}/status": {
            "patch": {
                "tags": [
                    "Seller - Requests"
                ],
                "summary": "Accept or reject a purchase request",
                "operationId": "8d55a6b5f086ebc437a6fd174d831245",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "accepted",
                                            "rejected"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/seller/incoming-rental-requests": {
            "get": {
                "tags": [
                    "Seller - Requests"
                ],
                "summary": "List rental requests for seller's cars",
                "operationId": "0b2c3050e337bad9149cdc0d44feb2ee",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/seller/rental-requests/{id}/status": {
            "patch": {
                "tags": [
                    "Seller - Requests"
                ],
                "summary": "Accept or reject a rental request",
                "operationId": "d13584d554c388edd3e23de1855118f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "accepted",
                                            "rejected"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}